refactor(exec): split exec.ts module
